deepEqual added
