+++ index.js
@@ -1,43 +1,22 @@
-function tree(obj) {
-  let indent = -1;
+import { deepEqual } from "./modules/deepEqual.js";
 
-  function getName(obj) {
-    let str = "";
-
-    for (let key in obj) {
-      if (key === "name") {
-        str += obj[key];
-      } else if (key === "items") {
-        indent += 1;
-
-        for (let item of obj[key]) {
-          str += `\n${
-            indent > 0 && indent < obj[key].length ? "|" : ""
-          }${" ".repeat(indent)}|__`;
-          str += `${getName(item)}`;
-        }
-
-        indent = 0;
-      }
-    }
-    return str;
-  }
-
-  console.log(getName(obj));
-}
-
-const obj = {
-  name: 1,
-  items: [
-    {
-      name: 2,
-      items: [{ name: 3 }, { name: 4 }],
-    },
-    {
-      name: 5,
-      items: [{ name: 6 }],
-    },
-  ],
+const obj1 = {
+  a: {
+    b: 1,
+  },
 };
 
-tree(obj);
+const obj2 = {
+  a: {
+    b: 2,
+  },
+};
+const obj3 = {
+  a: {
+    b: 1,
+  },
+};
+
+deepEqual(obj1, obj1)
+deepEqual(obj1, obj2)
+deepEqual(obj1, obj3)
